add internal sweepCoef property to Shapes defining the coeffecient used to scale with angular velocity magnitude to bias conservative advancement. Computed the difference between max and minimum positive radius (so things like circle at centre have 0 bias, and square has much smaller bias than a thin rectangle). In this manner we improve performance without permitting tunelling.
Fixed some bugs, and added internal SweepDistance::staticSweep performing
a conservative advancement of the shape of one dynamic body, against
shape of a static body.

To this effect added sweepIntegrate and sweepVerify for optimised
variants of updatePos() from Space on a single body making use of
delta_rot to reduce number of calls to Math.sin/cos. and sweepVerify
performing optimised recalculation of worldCOM (circle) and world
vertices and axi (Polygon).