(release ready)
Improvement to CCD collision detection to deal with slight overshooting
due to thresholding and numerical innacuracies. If it is determined
that objects to not continously sweep against eachother, it does one
final check at curTOI = 1 to ensure that it hasn't accidently missed
anything.

Helps avoid even further the 1-frame delay in CCD bouncing.